Revert "Updates Submitter w/ Oracle fallback (#1671)"
This reverts commit 530cf510712abdbd83ebd442ea4c60d6c898a8ca.
