+++ ethergo/submitter/chain_queue.go
@@ -2,18 +2,15 @@ package submitter
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
-	"github.com/lmittmann/w3"
 	"github.com/lmittmann/w3/module/eth"
 	"github.com/lmittmann/w3/w3types"
 	"github.com/synapsecns/sanguine/core"
 	"github.com/synapsecns/sanguine/core/metrics"
 	"github.com/synapsecns/sanguine/ethergo/client"
 	"github.com/synapsecns/sanguine/ethergo/submitter/db"
-	"github.com/synapsecns/sanguine/ethergo/util"
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
 	"golang.org/x/sync/errgroup"
@@ -47,10 +44,6 @@ func (c *chainQueue) chainIDInt() int {
 }
 
 func (t *txSubmitterImpl) chainPendingQueue(parentCtx context.Context, chainID *big.Int, txes []db.TX) (err error) {
-	if len(txes) == 0 {
-		return nil
-	}
-
 	ctx, span := t.metrics.Tracer().Start(parentCtx, "submitter.ChainQueue")
 	defer func() {
 		metrics.EndSpanWithErr(span, err)
@@ -92,41 +85,30 @@ func (t *txSubmitterImpl) chainPendingQueue(parentCtx context.Context, chainID *
 
 		cq.bumpTX(gCtx, tx)
 	}
-	cq.updateOldTxStatuses(gCtx)
+    cq.updateOldTxStatuses(gCtx)
 
 	err = cq.g.Wait()
 	if err != nil {
 		return fmt.Errorf("error in chainPendingQueue: %w", err)
 	}
 
-	cq.storeAndSubmit(ctx, span)
+	sort.Slice(cq.reprocessQueue, func(i, j int) bool {
+		return cq.reprocessQueue[i].Nonce() < cq.reprocessQueue[j].Nonce()
+	})
+
+	calls := make([]w3types.Caller, len(cq.reprocessQueue))
+	txHashes := make([]common.Hash, len(cq.reprocessQueue))
+	for i, tx := range cq.reprocessQueue {
+		calls[i] = eth.SendTx(tx.Transaction).Returns(&txHashes[i])
+	}
+
+	cq.storeAndSubmit(ctx, calls, span)
 
 	return nil
 }
 
 // storeAndSubmit stores the txes in the database and submits them to the chain.
-func (c *chainQueue) storeAndSubmit(ctx context.Context, parentSpan trace.Span) {
-	sort.Slice(c.reprocessQueue, func(i, j int) bool {
-		return c.reprocessQueue[i].Nonce() < c.reprocessQueue[j].Nonce()
-	})
-
-	calls := make([]w3types.Caller, len(c.reprocessQueue))
-	txHashes := make([]common.Hash, len(c.reprocessQueue))
-	spanners := make([]trace.Span, len(c.reprocessQueue))
-
-	for i, tx := range c.reprocessQueue {
-		_, span := c.metrics.Tracer().Start(ctx, "chainPendingQueue.submit", trace.WithAttributes(util.TxToAttributes(tx.Transaction)...))
-		spanners[i] = span
-		defer func() {
-			// in case this span doesn't get ended automatically below, end it here.
-			if span.IsRecording() {
-				metrics.EndSpan(span)
-			}
-		}()
-
-		calls[i] = eth.SendTx(tx.Transaction).Returns(&txHashes[i])
-	}
-
+func (c *chainQueue) storeAndSubmit(ctx context.Context, calls []w3types.Caller, span trace.Span) {
 	var wg sync.WaitGroup
 	wg.Add(2)
 
@@ -136,38 +118,25 @@ func (c *chainQueue) storeAndSubmit(ctx context.Context, parentSpan trace.Span) 
 		defer wg.Done()
 		err := c.db.PutTXS(storeCtx, c.reprocessQueue...)
 		if err != nil {
-			parentSpan.AddEvent("could not store txes", trace.WithAttributes(attribute.String("error", err.Error())))
+			span.AddEvent("could not store txes", trace.WithAttributes(attribute.String("error", err.Error())))
 		}
 	}()
 
 	go func() {
 		defer wg.Done()
 		err := c.client.BatchWithContext(ctx, calls...)
-
-		// Optimistically store all transactions as pending in the first goroutine.
-		// The call is made concurrently in the second goroutine.
-		// If the second goroutine finishes first, updating the status,
-		// there's no need to store them as pending, hence the context is canceled early.
 		cancelStore()
 		for i := range c.reprocessQueue {
-			span := spanners[i]
 			if err != nil {
 				c.reprocessQueue[i].Status = db.FailedSubmit
-				//nolint: errorlint, forcetypeassert
-				if callErrs, ok := err.(w3.CallErrors); ok && len(callErrs) > i {
-					span.RecordError(errors.New(callErrs[i].Error()))
-				} else if err != nil {
-					span.RecordError(err)
-				}
 			} else {
 				c.reprocessQueue[i].Status = db.Submitted
 			}
-			span.End()
 		}
 
 		err = c.db.PutTXS(ctx, c.reprocessQueue...)
 		if err != nil {
-			parentSpan.AddEvent("could not store txes", trace.WithAttributes(attribute.String("error", err.Error())))
+			span.AddEvent("could not store txes", trace.WithAttributes(attribute.String("error", err.Error())))
 		}
 	}()
 	wg.Wait()
@@ -240,7 +209,7 @@ func (c *chainQueue) bumpTX(parentCtx context.Context, ogTx db.TX) {
 			return fmt.Errorf("could not sign tx: %w", err)
 		}
 
-		span.AddEvent("add to reprocess queue", trace.WithAttributes(util.TxToAttributes(tx)...))
+		span.AddEvent("add to reprocess queue", trace.WithAttributes(txToAttributes(tx)...))
 
 		c.addToReprocessQueue(db.TX{
 			Transaction: tx,
